requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=h2bbff1b_5
ca-certificates=2024.3.11=haa95532_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
expat=2.6.2=hd77b12b_0
fonttools=4.51.0=pypi_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_0
matplotlib=3.8.4=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=h2bbff1b_0
packaging=24.0=pypi_0
pillow=10.3.0=pypi_0
pip=23.3.1=py312haa95532_0
pyparsing=3.1.2=pypi_0
python=3.12.3=h1d929f7_0
python-dateutil=2.9.0.post0=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py312haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
tzdata=2024a=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.41.2=py312haa95532_0
xz=5.4.6=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
